+++ src/fastapi_test_task/api/user_api.py
@@ -3,6 +3,7 @@ from typing import List
 
 from repositories.user_repository import UserRepository, UserRepositoryCreate, UserRepositoryUpdate
 from services.user_service import UserService
+from api.auth import oauth2_scheme, is_token_valid
 
 
 router = APIRouter(
@@ -11,38 +12,50 @@ router = APIRouter(
 
 
 @router.get('/', response_model=List[UserRepository])
-def get_users(service: UserService = Depends()):
-    return service.get_all()
+def get_users(
+    service: UserService = Depends(),
+    token: str = Depends(oauth2_scheme)
+):
+    if is_token_valid(token):
+        return service.get_all()
 
 
 @router.get('/{user_id}', response_model=UserRepository)
 def get_user(
     user_id: int,
     service: UserService = Depends(),
+    token: str = Depends(oauth2_scheme)
 ):
-    return service.get(user_id)
+    if is_token_valid(token):
+        return service.get(user_id)
 
 
 @router.post('/', response_model=UserRepository)
 def create_user(
     user_data: UserRepositoryCreate,
-    service: UserService = Depends()
+    service: UserService = Depends(),
+    token: str = Depends(oauth2_scheme)
 ):
-    return service.create(user_data)
+    if is_token_valid(token):
+        return service.create(user_data)
 
 
 @router.put('/{user_id}', response_model=UserRepository)
 def update_user(
     user_id: int,
     user_data: UserRepositoryUpdate,
-    service: UserService = Depends()
+    service: UserService = Depends(),
+    token: str = Depends(oauth2_scheme)
 ):
-    return service.update(user_id, user_data)
+    if is_token_valid(token):
+        return service.update(user_id, user_data)
     
 
 @router.delete('/{user_id}')
 def delete_user(
     user_id: int,
     service: UserService = Depends(),
+    token: str = Depends(oauth2_scheme)
 ):
-    return service.delete(user_id)
+    if is_token_valid(token):
+        return service.delete(user_id)
